Generalize credentials of vault
