+++ bin/generic/credential/certificate/add/examples.md
@@ -1,0 +1,3 @@
+```bash
+{{command_name}} --{{base_type}} my-{{base_type}} --name my-key --sshkey my-home-ssh
+```
