+++ bin/generic/credential/password/add/examples.md
@@ -1,0 +1,3 @@
+```bash
+{{command_name}} --name my-pass --password secret-password --{{base_type}} my-{{base_type}}
+```
