--- bin/vault/credential/certificate/add/examples.md
@@ -1,3 +1,0 @@
-```bash
-{{command_name}} --vault my-vault --name my-key --sshkey my-home-ssh
-```
