--- bin/vault/credential/password/add/examples.md
@@ -1,3 +1,0 @@
-```bash
-{{command_name}} --name my-pass --password secret-password --vault my-vault
-```
