+++ bin/vault/tests.js
@@ -108,20 +108,7 @@ ava.serial('vault recreate from snapshot', async t => {
 
 ava.serial('vault credential password life cycle', async t => {
     const vault = await tests.run(`vault create --name vault-test-${now} --size 10`);
-    const name = `vault-password-${now}`;
-    const secret = await tests.getToken();
-
-    await tests.run(`vault credential password add --name ${name} --password ${secret} --vault ${vault._id}`);
-    const list = await tests.run(`vault credential password list --vault ${vault._id}`);
-    t.true(list.some(p => p.name === name));
-
-    const password = await tests.run(`vault credential password show --vault ${vault._id} --password ${name}`);
-    await tests.run(`vault credential password rename --vault ${vault._id} --password ${name} --new-name secure-password`);
-    await tests.run(`vault credential password show --vault ${vault._id} --password ${password._id}`);
-
-    await tests.remove('vault credential password', password, {
-        deleteParams: `--vault ${vault._id}`,
-    });
+    await tests.passwordLifeCycle(t, 'vault', vault);
     await tests.remove('vault', vault);
 });
